Ostad Assignment 03 Completed

+++ src/shapes/Square.java
@@ -2,9 +2,9 @@ package shapes;
 
 public class Square extends Rectangle {
 
-	public Square(int side) {
+	public Square(double side) {
 		super(side, side);
-		
+
 	}
 
 }
